Add files via upload
Welcome to my repository where you can find the code for the Bangalore house prediction model. This project aims to predict the prices of houses in Bangalore, India, based on various factors such as location, size, and amenities.

The model uses machine learning algorithms to analyze a dataset of house prices and their corresponding features, and then predicts the price of a new house based on those features. The code is written in Python and utilizes popular machine learning libraries such as Scikit-learn, Pandas, and Numpy.

In this repository, you will find the source code for the model, along with the necessary datasets and pre-trained models. The code is well-documented and can be easily understood and modified for your specific needs. Additionally, you will find detailed instructions on how to run the model and interpret the results.

Whether you're a data scientist looking to learn more about machine learning, or a real estate agent looking to better understand the Bangalore housing market, this repository will provide you with a valuable tool for predicting house prices.